+++ Caffe-Mana/Domains/CacheScene.swift
@@ -1,0 +1,8 @@
+//
+//  CacheScene.swift
+//  Caffe-Mana
+//
+//  Created by 観 on 2023/05/20.
+//
+
+import Foundation
